+++ packages/devui-vue/devui/table/src/components/body/body.tsx
@@ -1,8 +1,7 @@
-import { defineComponent, inject, computed, PropType, toRef } from 'vue';
+import { defineComponent, inject, computed } from 'vue';
 import { TABLE_TOKEN } from '../../table-types';
-import { Column } from '../column/column-types';
+import TD from '../body-td/body-td';
 import { Checkbox } from '../../../../checkbox';
-import { useFixedColumn } from '../../composable/use-table';
 import { useNamespace } from '../../../../shared/hooks/use-namespace';
 import './body.scss';
 
@@ -12,19 +11,9 @@ export default defineComponent({
     const table = inject(TABLE_TOKEN);
     const { _data: data, _columns: columns, _checkList: checkList, isFixedLeft } = table.store.states;
     const ns = useNamespace('table');
-
-    // 移动到行上是否高亮
     const hoverEnabled = computed(() => table.props.rowHoveredHighlight);
 
-    // 行前的 checkbox
-    const tdAttrs = computed(() =>
-      isFixedLeft.value
-        ? {
-            class: `${ns.m('sticky-cell')} left`,
-            style: 'left:0;',
-          }
-        : null
-    );
+    const tdAttrs = computed(() => (isFixedLeft.value ? { class: `${ns.m('sticky-cell')} left`, style: 'left:0;' } : null));
     const renderCheckbox = (index: number) =>
       table.props.checkable ? (
         <td {...tdAttrs.value}>
@@ -45,32 +34,6 @@ export default defineComponent({
           );
         })}
       </tbody>
-    );
-  },
-});
-
-const TD = defineComponent({
-  props: {
-    column: {
-      type: Object as PropType<Column>,
-    },
-    row: {
-      type: Object,
-    },
-    index: {
-      type: Number,
-    },
-  },
-  setup(props: { column: Column; row: any; index: number }) {
-    const column = toRef(props, 'column');
-
-    // 固定列
-    const { stickyCell, offsetStyle } = useFixedColumn(column);
-
-    return () => (
-      <td class={stickyCell.value} style={offsetStyle.value}>
-        {column.value.renderCell(props.row, props.index)}
-      </td>
     );
   },
 });
